requirements.txt
numpy
opencv-python
Pillow
matplotlib
tqdm